+++ branches/avidemux_2.1_branch/avidemux/ADM_script/ADM_JSAvidemuxVideo.cpp
@@ -33,7 +33,7 @@ extern VF_FILTERS filterGetTagFromName(char *inname);
 //extern uint8_t indexMpeg(char *mpeg,char *file,uint8_t aid);
 extern uint8_t A_ListAllBlackFrames( char *file );
 extern uint8_t loadVideoCodecConfString( char *name);
-extern int ADM_saveRaw (char *name);
+extern uint8_t ADM_saveRaw (char *name);
 extern int A_saveJpg (char *name);
 extern uint8_t loadVideoCodecConf( char *name);
 extern int videoCodecConfigure(char *p,uint32_t i, uint8_t  *c);
